+++ ruoyi-system/src/main/java/com/ruoyi/system/service/impl/MeetingServiceImpl.java
@@ -6,9 +6,11 @@ import java.util.HashMap;
 import java.util.List;
 
 import com.ruoyi.system.mapper.PurchaseMapper;
+import org.activiti.engine.HistoryService;
 import org.activiti.engine.IdentityService;
 import org.activiti.engine.RuntimeService;
 import org.activiti.engine.TaskService;
+import org.activiti.engine.history.HistoricProcessInstance;
 import org.activiti.engine.runtime.ProcessInstance;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
@@ -44,6 +46,9 @@ public class MeetingServiceImpl implements IMeetingService
 
     @Resource
     IdentityService identityService;
+
+    @Resource
+    HistoryService historyService;
 
     /**
      * 查询会议
@@ -113,10 +118,13 @@ public class MeetingServiceImpl implements IMeetingService
         String[] keys = Convert.toStrArray(ids);
         for (String key : keys) {
             ProcessInstance process = runtimeService.createProcessInstanceQuery().processDefinitionKey("meeting").processInstanceBusinessKey(key).singleResult();
-            try {
-                runtimeService.deleteProcessInstance(process.getId(),"删除");
-            } catch (Exception e) {
-                e.printStackTrace();
+            if (process != null) {
+                runtimeService.deleteProcessInstance(process.getId(), "删除");
+            }
+            // 删除历史数据
+            HistoricProcessInstance history = historyService.createHistoricProcessInstanceQuery().processDefinitionKey("meeting").processInstanceBusinessKey(key).singleResult();
+            if (history != null) {
+                historyService.deleteHistoricProcessInstance(history.getId());
             }
             meetingMapper.deleteMeetingByIds(Convert.toStrArray(ids));
         }
